Separate TestingSceneDelegate
