+++ WooCommerce/WooCommerceTests/System/TestingAppDelegate.swift
@@ -35,20 +35,3 @@ class TestingAppDelegate: AppDelegate {
         return config
     }
 }
-
-// MARK: - Testing Scene Delegate (UIWindowSceneDelegate)
-final class TestingSceneDelegate: UIResponder, UIWindowSceneDelegate {
-    var window: UIWindow?
-
-    func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
-        guard let windowScene = scene as? UIWindowScene else { return }
-
-        let bundle = Bundle(for: TestingAppDelegate.self)
-        let storyboard = UIStoryboard(name: "TestingMode", bundle: bundle)
-
-        let window = UIWindow(windowScene: windowScene)
-        window.rootViewController = storyboard.instantiateInitialViewController()
-        window.makeKeyAndVisible()
-        self.window = window
-    }
-}
