+++ WooCommerce/WooCommerceTests/System/TestingSceneDelegate.swift
@@ -1,0 +1,17 @@
+import UIKit
+
+final class TestingSceneDelegate: UIResponder, UIWindowSceneDelegate {
+    var window: UIWindow?
+
+    func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
+        guard let windowScene = scene as? UIWindowScene else { return }
+
+        let bundle = Bundle(for: TestingAppDelegate.self)
+        let storyboard = UIStoryboard(name: "TestingMode", bundle: bundle)
+
+        let window = UIWindow(windowScene: windowScene)
+        window.rootViewController = storyboard.instantiateInitialViewController()
+        window.makeKeyAndVisible()
+        self.window = window
+    }
+}
